fix: list when there's an error would blow up
Since `fns` response is `null` when there's an error in the `lambda.listFunctions`, the line `fns.Functions.map` blows up, since `Functions` prop is undefined (on `null`). So let's pass the error back early if it's there.

Note that I've not gone through all the calls to check this pattern exists, I just know that it causes problems both on `npm run list` and `npm run deploy ...`.